Create homework4.js
Assignment: Create a Ghost object in javascript. In the constructor, pass in the initial position, and the color of the ghost (It's one of those Pac-man ghosts). Give it methods to move up, down, left, right. Make a toString method which prints out relevant information about the ghost. Like in Pac-man, the ghost should be able to wrap around, so once it leaves the "screen" (which has a 16x16 resolution), it should appear on the other side.